requirements.txt
tensorflow==2.19.0
numpy>=1.25
opencv-python
matplotlib
Pillow